adds support for chats with toolcalling
NOTES
we noted that Microsoft's phi4 model returns an object that does
not conform to Ollama's API documentation.

So we tried qwen3:0.6b (qwen2 does not support tool calling) with
success. We still need to add more code but at least the initial
response has the expected tool_calls field (which is what
Microsoft's phi4-mini model was lacking).