impls toolcalling with history
NOTES
the test code passes the requirement of generating a complete response
that requires the chatbot to complete a task by calling a tool

in this case the tool is used to generate a lammps input file for
a molecular dynamics simulation

in the final response the model claims that the lammps input file
is properly formatted and that the simulation parameters make sense
(runtime length for instance) given the temperature and pressure

we opted to create a map to `map` the tool name to the actual tool
(function) for performance; the implementation of tool() still needs
to be worked on in order to achieve our performance goal (O(1)
search time complexity).

FINAL
for this to work, we need to strip/remove the tool_calls field
from the data sent to the llm so that it won't try to call a tool
and this is why we do this prior to getting the final response; if
we don't model will think that we still want to use a tool and
it returns a tool_calls field (which is not what we want). I have
verified the ollama documentation and sample codes and this is how
things are done so we are confindent in our work.